change invoice smallint->int

+++ laravel-snregister/database/migrations/2017_10_04_135541_create_serial_table.php
@@ -16,7 +16,7 @@ class CreateSerialTable extends Migration
         Schema::create('Serial_table', function (Blueprint $table) {
             $table->increments('id');
 			$table->string('company');
-			$table->smallInteger('invoice');
+			$table->integer('invoice');
 			$table->string('product');
 			$table->string('product_sn');
 			$table->string('hdd');
